Clean up exports
