+++ types/sabaki__immutable-gametree/src/GameTree.d.ts
@@ -1,5 +1,87 @@
-import { NodeObject, Primitive } from "../index";
 import Draft = require("./Draft");
+type Primitive = string | number | symbol;
+
+type Property =
+    | "AB"
+    | "AE"
+    | "AN"
+    | "AP"
+    | "AR"
+    | "AS"
+    | "AW"
+    | "B"
+    | "BL"
+    | "BM"
+    | "BR"
+    | "BT"
+    | "C"
+    | "CA"
+    | "CP"
+    | "CR"
+    | "DD"
+    | "DM"
+    | "DO"
+    | "DT"
+    | "EV"
+    | "FF"
+    | "FG"
+    | "GB"
+    | "GC"
+    | "GM"
+    | "GN"
+    | "GW"
+    | "HA"
+    | "HO"
+    | "IP"
+    | "IT"
+    | "IY"
+    | "KM"
+    | "KO"
+    | "LB"
+    | "LN"
+    | "MA"
+    | "MN"
+    | "N"
+    | "OB"
+    | "ON"
+    | "OT"
+    | "OW"
+    | "PB"
+    | "PC"
+    | "PL"
+    | "PM"
+    | "PW"
+    | "RE"
+    | "RO"
+    | "RU"
+    | "SE"
+    | "SL"
+    | "SO"
+    | "SQ"
+    | "ST"
+    | "SU"
+    | "SZ"
+    | "TB"
+    | "TE"
+    | "TM"
+    | "TR"
+    | "TW"
+    | "UC"
+    | "US"
+    | "V"
+    | "VW"
+    | "W"
+    | "WL"
+    | "WR"
+    | "WT";
+
+
+interface NodeObject<ID extends Primitive = number> {
+    id: ID;
+    data: Partial<Record<Property, string[]>>;
+    parentId: ID | null;
+    children: NodeObject<ID>[];
+}
 
 interface GameTreeOptions<ID extends Primitive = number> {
     getId?: () => ID;
@@ -57,4 +139,7 @@ declare class GameTree<ID extends Primitive = number> {
     toJSON(): NodeObject<ID>;
 }
 
+declare namespace GameTree {
+    export { Draft, NodeObject , Primitive, Property};
+}
 export = GameTree;
